pyproject.toml
[project]
name = "savemate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "spotdl>=4.2.11",
    "yt-dlp>=2025.4.30",
]
